fix: new telegram users can use the bot
Telegram has been moved to "int64" for user's id, but the bot's database
has still "int32" for user's id.

+++ app/models/user.py
@@ -4,7 +4,7 @@ from app.models.db import TimedBaseModel, db
 class User(TimedBaseModel):
     __tablename__ = "users"
 
-    user_id = db.Column(db.Integer, primary_key=True, index=True, unique=True)
+    user_id = db.Column(db.BigInteger, primary_key=True, index=True, unique=True)
     grade = db.Column(db.Integer)
     subject = db.Column(db.String)
     author = db.Column(db.String)
